backend/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
pymongo==4.0.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
